Implement F&A in authors' panel
